+++ src/index.ts
@@ -1,3 +1,6 @@
 import "dotenv/config"
 
 export { createPlugin } from "./plugin"
+
+export { extractResourceLocator } from './utils/extract-resource-locator'
+export { extractResourceMapper } from './utils/extract-resource-mapper'